+++ unreal/Plugins/RaftSim/Source/RaftSimUI/RaftSimUI.Build.cs
@@ -5,6 +5,6 @@ public class RaftSimUI : ModuleRules
     public RaftSimUI(ReadOnlyTargetRules Target) : base(Target)
     {
         PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
-        PublicDependencyModuleNames.AddRange(new[] { "Core", "CoreUObject", "Engine", "UMG", "RaftSimCore" });
+        PublicDependencyModuleNames.AddRange(new[] { "Core", "CoreUObject", "Engine", "UMG", "Slate", "SlateCore", "RaftSimCore" });
     }
 }